Add unit tests for `CLI` class
Add unit tests for `CLI` class, however it is not easy to process an
inactive Stream, without wrapping it a Future, returning the Future to a
Completer, forcing the Future to be complete and then inspect the
results. Instead, I have chosen to call a remote process, local to the
system, which calls the main application and pretends to pass it filename
or User Input's via `stdin`.

== References:
- [Process class \- dart:io library \- Dart API]
  (https://api.dartlang.org/stable/1.24.3/dart-io/Process-class.html)